% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepcombat_dataloader.R
\name{create_model_matrices_from_train}
\alias{create_model_matrices_from_train}
\title{Create model matrices for batch correction from trained model}
\usage{
create_model_matrices_from_train(
  setup_input,
  covariate_formula,
  batch_formula,
  feature_matrix,
  covariate_matrix,
  reference_batch = NULL
)
}
\arguments{
\item{setup_input}{output from deepcombat_setup containing training estimands and settings}

\item{covariate_formula}{formula specifying the covariates to preserve. Ex: ~ Covariate1 + Covariate2}

\item{batch_formula}{formula specifying the batch variable to correct for. Ex: ~ Batch}

\item{feature_matrix}{a matrix or dataframe of features with rows representing subjects and columns representing features.}

\item{covariate_matrix}{a matrix or dataframe of covariates with rows representing subjects and columns representing covariates. Columns must be named such that covariate_formula can extract the correct columns.}

\item{reference_batch}{the name of the batch to be used as the reference level for batch correction. If not specified,
the function will generate a target batch that maps to an intermediate space between all batches. Intermediate-space harmonization is currently only implemented for correction of two batches.}
}
\value{
a list with the following components:
\describe{
\item{features}{a scaled feature matrix}
\item{covariates}{the model matrix for covariates, excluding intercept, with each column scaled to have values between 0 and 1.}
\item{batch}{the model matrix for batch, excluding intercept}
\item{target_batch}{the model matrix for the target batch, which will be used to adjust the batch effect in the feature matrix}
}
}
\description{
This function takes in a formula for covariates, a formula for batch, a feature matrix,
and a covariate matrix and generates the corresponding model matrices for these variables, excluding intercept.
It also generates a target batch model matrix that will be used to adjust the batch effect in the feature matrix.
}
\examples{
\dontrun{
create_model_matrices_from_train(setup_input, covariate_formula = ~ age + sex + diagnosis,
batch_formula = ~ batch,
feature_matrix = data, covariate_matrix = covariates,
reference_batch = "batch1")
}
}
