% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepcombat_architecture.R
\name{deepcombat_cvae}
\alias{deepcombat_cvae}
\title{DeepComBat CVAE architecture}
\usage{
deepcombat_cvae(feature_dim, latent_dim, n_hidden = 3, n_batch, n_covariate)
}
\arguments{
\item{feature_dim}{Number of features to harmonize}

\item{latent_dim}{Desired latent space size}

\item{n_hidden}{Number of hidden layers}

\item{n_batch}{Number of batches}

\item{n_covariate}{Number of covariates}
}
\value{
nn_module object with desired architecture.
}
\description{
Code for CVAE architecture. This function is called through helper functions of deepcombat_setup, deepcombat_trainer and deepcombat_harmonize.
}
