% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepcombat_trainer.R
\name{deepcombat_harmonize}
\alias{deepcombat_harmonize}
\title{DeepComBat Harmonization}
\usage{
deepcombat_harmonize(
  setup,
  trained_model,
  plot_logvars = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{setup}{A deepcombat_setup_object.}

\item{trained_model}{A trained DeepComBat CVAE model from deepcombat_trainer.}

\item{plot_logvars}{A logical value indicating whether to plot the density of logvars according to dimension -- this may be useful for manual hyperparameter tuning. Default is FALSE.}

\item{verbose}{A logical value indicating whether to print progress messages. Default is FALSE.}
}
\value{
A deepcombat_harmonized_object containing the harmonized data, logvars, and additional outputs that may be useful for methodologists/advanced users.
}
\description{
This function performs harmonization using DeepComBat. It takes in a setup object and a trained DeepComBat CVAE model, and returns the harmonized output.
}
\examples{
\dontrun{
harmonized_output <- deepcombat_harmonize(setup_obj, trained_model, 
plot_logvars = TRUE, verbose = TRUE)
}

}
